requirements.txt
# Core ML and Data Science
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# NLP and Text Processing
nltk>=3.6.0
spacy>=3.4.0
textblob>=0.15.0

# Deep Learning (if needed for advanced models)
tensorflow>=2.8.0
torch>=1.10.0

# Web Framework
streamlit>=1.20.0

# API and HTTP
requests>=2.25.0
httpx>=0.24.0

# Environment and Configuration
python-dotenv>=0.19.0

# LangChain and LLM Integration
langchain>=0.1.0
langchain-core>=0.1.0
langchain-groq>=0.0.1
langchain-community>=0.0.10

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Jupyter (for development)
jupyter>=1.0.0
ipykernel>=6.0.0

# Testing
pytest>=6.0.0
pytest-cov>=3.0.0

# Code Quality
black>=22.0.0
flake8>=4.0.0
isort>=5.0.0

# Utilities
tqdm>=4.62.0
click>=8.0.0
rich>=12.0.0

# Optional: Model Serialization
joblib>=1.1.0
pickle5>=0.0.11

# Optional: Advanced NLP
transformers>=4.20.0
tokenizers>=0.12.0

# Optional: Data Validation
pydantic>=1.9.0

# Optional: Async Support
aiohttp>=3.8.0
asyncio>=3.4.3 